request-header/pyproject.toml
[tool.poetry]
name = "request-header"
version = "0.1.0"
description = "Request Header Parser Microservice from FreeCodeCamp BEDAAPIs certification."
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.9"
djangorestframework = "^3.12.4"
django-cors-headers = "^3.11.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = "^22.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
